requirements.txt
sco-datastore>=0.5.0
